304: Move budget summary stuff into own sub-folder
