+++ frontend/src/pages/cans/detail/budgetSummary/budgetSummary.css
